requirements.txt
pysinric==0.1.1
six==1.15.0
wakeonlan==1.1.6
websocket-client==0.57.0
